Renames tiltakshaver table to generic part table
Refactors the `custom-table-tiltakshaver` component to a more generic `custom-table-part` component.

This change allows the component to be reused for displaying different types of "part" data, not just "tiltakshaver".

The component tag name, related files, and documentation have been updated to reflect the new name and purpose.

Adds a `partType` property to specify the type of part represented in the table.
