+++ src/components/data-components/custom-table-part/index.js
@@ -1,0 +1,30 @@
+// Global functions
+import { renderFeedbackListElement } from "../../../functions/feedbackHelpers.js";
+import { getComponentContainerElement } from "../../../functions/helpers.js";
+import { instantiateComponent } from "../../../functions/componentHelpers.js";
+
+// Local functions
+import { renderPartTable } from "./renderers.js";
+
+// Stylesheet
+import "./styles.css" with { type: "css" };
+
+export default customElements.define(
+    "custom-table-part",
+    class extends HTMLElement {
+        async connectedCallback() {
+            const component = instantiateComponent(this);
+            const componentContainerElement = getComponentContainerElement(this);
+            if (component?.hideIfEmpty && component.isEmpty && !!componentContainerElement) {
+                componentContainerElement.style.display = "none";
+            } else {
+                const feebackListElement = component.hasValidationMessages && renderFeedbackListElement(component?.validationMessages);
+                const partTable = renderPartTable(component);
+                this.appendChild(partTable);
+                if (feebackListElement) {
+                    this.appendChild(feebackListElement);
+                }
+            }
+        }
+    }
+);
