+++ src/components/data-components/custom-table-part/styles.css
@@ -1,4 +1,4 @@
-custom-table-tiltakshaver {
+custom-table-part {
     display: block;
     width: 100%;
 }
