--- src/components/data-components/custom-table-tiltakshaver/textResourceBindings.js
@@ -1,9 +1,0 @@
-export default {
-    tiltakshaver: {
-        title: "resource.tiltakshaver.header",
-        "col-1": "resource.tiltakshaver.navn.title",
-        "col-2": "resource.tiltakshaver.telefonnummer.title",
-        "col-3": "resource.tiltakshaver.epost.title",
-        "emptyFieldText-default": "resource.emptyFieldText.default"
-    }
-};
